Move to emit_te
